+++ dart_sample/lib/sample_routes.dart
@@ -50,6 +50,7 @@ class SampleRoutes {
   static const String radioExpansionTile = '/radio-expansion-tile';
   static const String expansionPanel = '/expansion-panel';
   static const String stepper = '/stepper';
+  static const String about = '/about';
   static const String floatingActionButton = '/floating-action-button';
   static const String checkbox = '/checkbox';
   static const String switchDemo = '/switch';